// multi-step-form-react/src/enums/FormText.ts
export enum FormText {
  NameId = 'name',
  NameLabel = 'Name',
  NamePlaceholder = 'e.g. Stephen King',
  EmailId = 'email',
  EmailLabel = 'Email Address',
  EmailPlaceholder = 'e.g. stephenking@lorem.com',
  PhoneId = 'phone',
  PhoneLabel = 'Phone Number',
  PhonePlaceholder = 'e.g. +1 234 567 890',
}