// multi-step-form-react/src/enums/AppText.ts
export enum AppText {
  PersonalInfoDescription = 'Please provide your name, email, address, and phone number.',
  NamePlaceholder = 'e.g. Stephen King',
  EmailPlaceholder = 'e.g. stephenking@lorem.com',
  PhonePlaceholder = 'e.g. +1 234 567 890',
  FieldRequired = 'This field is required',
  InvalidPhone = 'Invalid phone number',
}